requirements.txt
numpy
pandas
Pillow
plotly
requests
streamlit
PyPortfolioOpt
datetime
matplotlib
